cmms-agent/requirements.txt
openai
requests
python-dotenv
playwright
fastapi
uvicorn[standard]
